+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 
 ##master
 - Added `autoload.php` for autolading files without composer
-- Added `React JSX` definition
+
+###Languages
+- React JSX
+
+##Plugins
+-  Highlight Keyword
 
 ##v0.1.0
 - Initial release